+++ WordPressKitTests/ActivityServiceRemoteTests.swift
@@ -168,7 +168,7 @@ class ActivityServiceRemoteTests: RemoteTestCase, RESTTestable {
                                   }, failure: { error in
                                       let error = error as NSError
                                       XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-                                      XCTAssertEqual(error.code, WordPressComRestApiError.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
+                                      XCTAssertEqual(error.code, WordPressComRestApiErrorCode.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
                                       expect.fulfill()
                                   })