+++ CHANGELOG.md
@@ -36,6 +36,7 @@ _None._
 
 - `WordPressComRestApi` initialisers now accept a `baseURL: URL` parameter instead of `baseUrlString: String`. [#691]
 - Removed the async functions in `WordPressComRestApi`. [#692]
+- `WordPressComRestApiError` is renamed to `WordPressRestApiErrorCode`, and no longer conforms to `Swift.Error`.
 
 ### New Features